+++ lab_01/src/validate_data.cpp
@@ -2,7 +2,7 @@
 #include "../inc/errors.h"
 #include "../inc/struct.h"
 
-int validate_connection(const connection_t &connection, const size_t cnt_points)
+int validate_connection(const connection_t &connection, const int cnt_points)
 {
     if (connection.index_dot_1 < cnt_points && connection.index_dot_2 < cnt_points)
         return OK;
